Python 2
I was assigned to code If the year was a leap year and I was able to use modulus and making sure when printing the answer It was a string. Also I was able to do a other project by using contact names and removing a desired name and organizing the list in order by the alphabet.